% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stepwise_regression.R
\name{comput_inverseR}
\alias{comput_inverseR}
\title{The function computes the inverse of a matrix.}
\usage{
comput_inverseR(V)
}
\arguments{
\item{V}{A square, symmetric, positive-definite matrix for which the inverse is to be computed.}
}
\value{
A list with the following elements:
\item{success}{A logical value indicating whether the inverse computation was successful.}
\item{inverse}{The computed inverse matrix, if successful. Returned as a matrix.}
\item{error}{An error message if both decompositions fail. Returned as a character string.}
}
\description{
This function attempts to compute the inverse of a matrix by first using Cholesky decomposition. 
If Cholesky decomposition fails, it falls back to LU decomposition. If both methods fail, it returns an error message.
}
