% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdSuSiE.R
\name{compute_b1b2}
\alias{compute_b1b2}
\title{The function computes the EB1 and EB2 matrices based on the provided input parameters.}
\usage{
compute_b1b2(alpha, mu1, mu2, column_config, nsex, nsnp)
}
\arguments{
\item{alpha}{Input alpha matrix.}

\item{mu1}{Input mu1 list.}

\item{mu2}{Input mu2 list.}

\item{column_config}{Configuration of columns.}

\item{nsex}{Number of sexes.}

\item{nsnp}{Number of SNPs.}
}
\value{
A list containing the computed EB1 and EB2 matrices.
}
\description{
The function computes the EB1 and EB2 matrices based on the provided input parameters.
}
