% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdSuSiE.R
\name{get_purity}
\alias{get_purity}
\title{The function computes the minimum, mean, and median of absolute correlations for specific positions.}
\usage{
get_purity(pos, Xcorr)
}
\arguments{
\item{pos}{A vector of positions to consider.}

\item{Xcorr}{A correlation matrix.}
}
\value{
A vector containing the minimum, mean, and median of the absolute correlations.
}
\description{
The function computes the minimum, mean, and median of absolute correlations for specific positions.
}
