% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdSuSiE.R
\name{gwas_plot_fun}
\alias{gwas_plot_fun}
\title{The function creates a GWAS or PIP plot for visualizing results from either univariate sex-dimorphic analysis or sdSuSiE.}
\usage{
gwas_plot_fun(data_plot, title_name, ylab_name, yintercept)
}
\arguments{
\item{data_plot}{A data frame containing results from sex-stratified GWASs or sdSuSiE.}

\item{title_name}{A string specifying the title of the plot.}

\item{ylab_name}{A string specifying the label for the y-axis, such as "-log10(p-value)" or "PIP".}

\item{yintercept}{A numeric value indicating the significance line, such as the genome-wide significance threshold.}
}
\value{
A ggplot object visualizing the results of either univariate sex-dimorphic analysis or sdSuSiE.
}
\description{
The function creates a GWAS or PIP plot for visualizing results from either univariate sex-dimorphic analysis or sdSuSiE.
}
