% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdSuSiE.R
\name{gwas_plot_fundouble}
\alias{gwas_plot_fundouble}
\title{The function creates GWAS plots for sex-stratified GWASs.}
\usage{
gwas_plot_fundouble(data_plot)
}
\arguments{
\item{data_plot}{A data frame containing results from sex-stratified GWASs.}
}
\value{
A ggplot object visualizing the results of sex-stratified GWASs.
}
\description{
The function creates GWAS plots for sex-stratified GWASs.
}
