% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdSuSiE.R
\name{in_CS}
\alias{in_CS}
\title{The function applies the in_CS_x function row-wise to a matrix to identify the SNPs that are included in the credible set for each row.}
\usage{
in_CS(res, coverage = 0.95)
}
\arguments{
\item{res}{A matrix where each row represents a set of SNP values.}

\item{coverage}{The desired coverage (default is 0.95).}
}
\value{
A matrix where each row is a binary vector indicating which SNPs are in the credible set.
}
\description{
The function applies the in_CS_x function row-wise to a matrix to identify the SNPs that are included in the credible set for each row.
}
