% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdSuSiE.R
\name{in_CS_x}
\alias{in_CS_x}
\title{The function identifies the SNPs that are included in the credible set to achieve the desired coverage.}
\usage{
in_CS_x(x, coverage = 0.95)
}
\arguments{
\item{x}{A numeric vector representing SNP values.}

\item{coverage}{The desired coverage (default is 0.95).}
}
\value{
A binary vector, where 1 indicates the SNP is in the credible set and 0 otherwise.
}
\description{
The function identifies the SNPs that are included in the credible set to achieve the desired coverage.
}
