% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdSuSiE.R
\name{n_in_CS_x}
\alias{n_in_CS_x}
\title{The function calculates the number of SNPs in a credible set required to achieve the desired coverage.}
\usage{
n_in_CS_x(x, coverage = 0.95)
}
\arguments{
\item{x}{A numeric vector representing SNP values.}

\item{coverage}{The desired coverage (default is 0.95 or 95%).}
}
\value{
The number of SNPs needed in the credible set.
}
\description{
The function calculates the number of SNPs in a credible set required to achieve the desired coverage.
}
