% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdSuSiE.R
\name{pre_optim}
\alias{pre_optim}
\title{The function prepares the input for the parameter space of the prior covariance matrix.}
\usage{
pre_optim(nsex, min_var, max_var)
}
\arguments{
\item{nsex}{Number of sexes.}

\item{min_var}{Lower bound of the prior variance.}

\item{max_var}{Upper bound of the prior variance.}
}
\value{
A list containing the initial parameters, lower bounds, upper bounds, diagonal element indices, and the number of sexes.
}
\description{
The function prepares the input for the parameter space of the prior covariance matrix.
}
