% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdSuSiE.R
\name{sdSuSiE}
\alias{sdSuSiE}
\title{The main function for sex-dimorphic mapping with sum of the single effects model (sdSuSiE).}
\usage{
sdSuSiE(
  R_mat_list,
  summary_stat_list,
  L = 10,
  residual_variance = NULL,
  prior_weights = NULL,
  config_weights = NULL,
  config_names = NULL,
  estimate_residual_variance = FALSE,
  cor_method = "min.abs.corr",
  cor_threshold = 0.5,
  max_iter = 100
)
}
\arguments{
\item{R_mat_list}{A list of length 2, where:
- The first element is the correlation matrix for males.
- The second element is the correlation matrix for females.
The column names of each correlation matrix must match the order of SNP names in the corresponding element of `summary_stat_list`.}

\item{summary_stat_list}{A list of length 2, where:
- The first element contains the summary statistics for males.
- The second element contains the summary statistics for females.
Each summary statistics element must include the following columns: SNP, Beta, Se, Z, and N. 
The order of SNPs in each summary statistics element must match the order of the corresponding correlation matrix in `R_mat_list`. 
The function `sdSuSiE` reconstructs sufficient statistics of each SNP for each sex using either:
- Marginal z-scores (Z) and sample sizes (N), or
- Marginal effect sizes (Beta) and standard errors (Se).}

\item{L}{The maximum number of non-zero effects assumed within the region (default: 10).}

\item{residual_variance}{A numeric vector (default: `c(1, 1)`) specifying the residual variance for males and females.}

\item{prior_weights}{A numeric vector of length p specifying the prior probability that each SNP has a non-zero effect. 
By default, prior weights are assumed to be equal for all SNPs.}

\item{config_weights}{A numeric vector of length 3 specifying the prior probabilities for different causal configurations:
- The first element represents the probability of a SNP having a common effect.
- The second element represents the probability of a SNP having a sex-dimorphic effect.
- The third element represents the probability of a SNP having both common and sex-dimorphic effects.
The three elements must sum to 1. The default prior values are `-1 + sqrt(2)`, `-1 + sqrt(2)`, and `(-1 + sqrt(2))^2`, respectively.}

\item{config_names}{A character vector of length 3 specifying the names for different causal configurations. The default prior values are `Common`, `Sex-dimorphic`, and `Both`, respectively.}

\item{estimate_residual_variance}{A logical value indicating whether the residual variance should be estimated (`TRUE`) or fixed (`FALSE`, default).}

\item{cor_method}{A string specifying the method to compute the 95% credible set based on correlations among SNPs in the region. Options include:
- `"min_abs_corr"`: Minimum absolute correlation (default).
- `"mean_abs_corr"`: Mean absolute correlation.
- `"median_abs_corr"`: Median absolute correlation.}

\item{cor_threshold}{A numeric value specifying the correlation threshold for `cor_method` (default: 0.5).}

\item{max_iter}{An integer specifying the maximum number of iterations to perform (default: 100).}
}
\value{
An R6 object with PIPs, credible sets, and other features of the sex-dimorphic fine-mapping result. 
\item{PIP}{A vector where each element represents the PIP of a SNP having a non-zero effect.}

\item{PIP_config}{A matrix where each column corresponds to the PIP of a causal configuration. The columns represent the PIP of a SNP having a common effect, a sex-dimorphic effect, or both.}

\item{alpha}{A list of length L, where each element is a matrix (p by the number of causal configurations) of PIPs. Each column represents the PIP for a specific scenario.}

\item{KL}{A vector containing the single-effect Kullback–Leibler (KL) divergences.}

\item{sigma2}{A numeric vector of estimated residual variance.}

\item{V}{A list of length L, where each element is a 2 by 2 matrix representing the prior variance estimate.}

\item{ELBO}{A vector storing the evidence lower bound achieved at each iteration of the model-fitting algorithm, which aims to maximize the ELBO.}

\item{Credible_Set}{The estimated credible sets for fine-mapping.}

\item{posterior_mean}{A matrix where the first column contains the posterior mean for \beta_c (common effects) and the second column contains the posterior mean for \eqn{\beta_d} (sex-dimorphic effects).}

\item{posterior_variance}{A matrix with the first column being the variance for \eqn{\beta_c}, the second column being the variance for \eqn{\beta_d}, and the third column being the covariance between \eqn{\beta_c} and \eqn{\beta_d}.}
}
\description{
sdSuSiE conduct sex-stratified GWAS fine-mapping to identify potentially sex-dimorphic causal SNPs associated with a complex trait of interest.
}
\examples{
library(sdSuSiE)
data(GWAS_M)
data(GWAS_F)
data(LD_M)
data(LD_F)
summary_list = list()
summary_list[[1]] = GWAS_M
summary_list[[2]] = GWAS_F
names(summary_list) = c("Male","Female")
LDmat_list = list()
LDmat_list[[1]] = LD_M
LDmat_list[[2]] = LD_F
names(LDmat_list) = c("Male","Female")
res <- sdSuSiE(LDmat_list, summary_list)
}
