% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdSuSiE.R
\name{sdSuSiE_get_cs}
\alias{sdSuSiE_get_cs}
\title{The function extracts credible sets from sdSuSiE results.}
\usage{
sdSuSiE_get_cs(
  res,
  Xcorr,
  coverage = 0.95,
  prior_tol = 1e-09,
  cor_method = cor_method,
  cor_threshold = cor_threshold
)
}
\arguments{
\item{res}{The result object from sdSuSiE.}

\item{Xcorr}{A list of correlation matrix used in sdSuSiE.}

\item{coverage}{A numeric value indicating the desired coverage for the credible set, default is \(0.95\).}

\item{prior_tol}{A numeric threshold value to determine prior variances, default is \(1e-9\).}

\item{cor_method}{A string indicating which method to use when computing purity, defalt is "min.abs.corr".}

\item{cor_threshold}{A numeric threshold value for purity measurement, default is 0.5.}
}
\value{
A list containing:
\item{cs}{The credible sets.}
\item{cs_category}{The category of each credible set having either common effect, sex-dimorphic effect, or both.}
\item{purity}{Purity statistics for each credible set.}
\item{cs_index}{The index for the credible sets.}
\item{coverage}{The achieved coverage for the credible sets.}
\item{requested_coverage}{The desired coverage (provided as input).}
}
\description{
The function extracts credible sets from sdSuSiE results.
}
