% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdSuSiE.R
\name{sdSuSiE_get_pip_config}
\alias{sdSuSiE_get_pip_config}
\title{The function extracts the maximum posterior inclusion probabilities (PIP) for each configuration from the given results.}
\usage{
sdSuSiE_get_pip_config(res)
}
\arguments{
\item{res}{The result object from sdSuSiE.}
}
\value{
A matrix containing the maximum PIP for each causal configuration.
}
\description{
The function extracts the maximum posterior inclusion probabilities (PIP) for each configuration from the given results.
}
