% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdSuSiE.R
\name{sdSuSiE_get_pip_either}
\alias{sdSuSiE_get_pip_either}
\title{The function extracts the maximum posterior inclusion probabilities (PIP) across all components from the given results.}
\usage{
sdSuSiE_get_pip_either(res)
}
\arguments{
\item{res}{The result object from sdSuSiE.}
}
\value{
A vector containing the maximum PIP for each SNP.
}
\description{
The function extracts the maximum posterior inclusion probabilities (PIP) across all components from the given results.
}
