% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdSuSiE.R
\name{sdSuSiEData}
\alias{sdSuSiEData}
\title{sdSuSiEData}
\value{
An updated `sdSuSiEData` class is used for handling and processing data in the sdSuSiE method.
}
\description{
An R6 class for managing and processing data relevant to the sdSuSiE approach.
}
\details{
This R6 class is designed for managing and processing data relevant to the sdSuSiE approach.
The class includes methods to initialize and process XtX and XtY data structures.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{R}}{A list of correlation matrix. Each element represents a correlation matrix for a specific group (e.g., males and females).}

\item{\code{Summary_Stat}}{A list of summary statistics. Each element contains the required columns, such as SNP, Beta, Se, Z, and N, for a specific group.}

\item{\code{var_y}}{Variance of y. Default is 1.}

\item{\code{Name_list}}{A list of SNP names extracted from the correlation matrices in `R`.}

\item{\code{N_sex}}{Number of sexs in `X`.}

\item{\code{XtX.diag}}{A vector containing the diagonal elements of the XtX matrix.}

\item{\code{XtX_list}}{A list of processed XtX matrices for each group.}

\item{\code{Xty_list}}{A list of processed XtY vectors for each group.}

\item{\code{N_list}}{A list of median sample sizes (`N`) for each group, computed from `Summary_Stat`.}

\item{\code{yty_list}}{A list of precomputed yty values for each group.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{R}}{A list of correlation matrix. Each element represents a correlation matrix for a specific group (e.g., males and females).}

\item{\code{Summary_Stat}}{A list of summary statistics. Each element contains the required columns, such as SNP, Beta, Se, Z, and N, for a specific group.}

\item{\code{var_y}}{Variance of y. Default is 1.}

\item{\code{Name_list}}{A list of SNP names extracted from the correlation matrices in `R`.}

\item{\code{N_sex}}{Number of sexs in `X`.}

\item{\code{XtX.diag}}{A vector containing the diagonal elements of the XtX matrix.}

\item{\code{XtX_list}}{A list of processed XtX matrices for each group.}

\item{\code{Xty_list}}{A list of processed XtY vectors for each group.}

\item{\code{N_list}}{A list of median sample sizes (`N`) for each group, computed from `Summary_Stat`.}

\item{\code{yty_list}}{A list of precomputed yty values for each group.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-sdSuSiEData-new}{\code{sdSuSiEData$new()}}
\item \href{#method-sdSuSiEData-XtX_diag}{\code{sdSuSiEData$XtX_diag()}}
\item \href{#method-sdSuSiEData-XtX_pro}{\code{sdSuSiEData$XtX_pro()}}
\item \href{#method-sdSuSiEData-Xty_pro}{\code{sdSuSiEData$Xty_pro()}}
\item \href{#method-sdSuSiEData-clone}{\code{sdSuSiEData$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sdSuSiEData-new"></a>}}
\if{latex}{\out{\hypertarget{method-sdSuSiEData-new}{}}}
\subsection{Method \code{new()}}{
Initialize the sdSuSiEData class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sdSuSiEData$new(X, Y, var_y = 1)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{X}}{A list or data frame containing the input correlation matrix.}

\item{\code{Y}}{A list or data frame containing the summary statistics.}

\item{\code{var_y}}{A numeric value representing the variance of Y, default is 1.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sdSuSiEData-XtX_diag"></a>}}
\if{latex}{\out{\hypertarget{method-sdSuSiEData-XtX_diag}{}}}
\subsection{Method \code{XtX_diag()}}{
Compute XtX.diag for each summary statistic
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sdSuSiEData$XtX_diag(Summary_Stat, Name_list)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{Summary_Stat}}{A list containing the summary statistics.}

\item{\code{Name_list}}{A list of names corresponding to the summary statistics.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A list of computed XtX.diag values for each summary statistic.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sdSuSiEData-XtX_pro"></a>}}
\if{latex}{\out{\hypertarget{method-sdSuSiEData-XtX_pro}{}}}
\subsection{Method \code{XtX_pro()}}{
Process XtX using the given R matrix and XtX.diag values
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sdSuSiEData$XtX_pro(R, XtX.diag, Name_list)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{R}}{A list or data frame representing the correlation matrix.}

\item{\code{XtX.diag}}{A list of XtX.diag values computed earlier.}

\item{\code{Name_list}}{A list of names corresponding to the summary statistics.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A list of processed XtX matrices for each summary statistic.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sdSuSiEData-Xty_pro"></a>}}
\if{latex}{\out{\hypertarget{method-sdSuSiEData-Xty_pro}{}}}
\subsection{Method \code{Xty_pro()}}{
Process XtY (product of XtX.diag and Beta for each summary statistic)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sdSuSiEData$Xty_pro(Summary_Stat, XtX.diag, Name_list)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{Summary_Stat}}{A list containing the summary statistics.}

\item{\code{XtX.diag}}{A list of XtX.diag values computed earlier.}

\item{\code{Name_list}}{A list of names corresponding to the summary statistics.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A list of computed XtY values for each summary statistic.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sdSuSiEData-clone"></a>}}
\if{latex}{\out{\hypertarget{method-sdSuSiEData-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sdSuSiEData$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
