% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdSuSiE.R
\name{sdSuSiEObject}
\alias{sdSuSiEObject}
\title{sdSuSiE Object}
\description{
An R6 object for handling the sdSuSiE analysis and storing its results.
}
\details{
This object is designed to manage and perform sdSuSiE analysis.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name_config}}{A character vector containing the names of causal configurations.}

\item{\code{column_config}}{A list specifying the possible configurations for columns.}

\item{\code{alpha}}{A list of alpha matrices, where each matrix corresponds to posterior inclusion probabilities for each component.}

\item{\code{mu1}}{A list of mu1 lists, where each list contains the posterior means for each causal configuration for each component.}

\item{\code{mu2}}{A list of mu2 lists, where each list contains the posterior variances for each causal configuration for each component.}

\item{\code{EB1}}{A list of EB1 matrices, where each matrix represents the posterior means of \eqn{\beta_c} and \eqn{\beta_d} for each component.}

\item{\code{EB2}}{A list of EB2 matrices, where each matrix represents the posterior variances for each component. The first column is the variance for \eqn{\beta_c}, the second column is the variance for \eqn{\beta_d}, and the third column is the covariance between \eqn{\beta_c} and \eqn{\beta_d}.}

\item{\code{Xr}}{A list of Xr matrices, representing the genetic part for each component.}

\item{\code{KL}}{A numeric vector containing the Kullback-Leibler (KL) divergence values for each component.}

\item{\code{lbf}}{A list of lbf matrices, where each matrix represents the log-Bayes factors for causal configuration for each component.}

\item{\code{ELBO}}{A numeric vector of Evidence Lower Bound (ELBO) values for each iteration, used to assess model convergence.}

\item{\code{sigma2}}{A numeric vector of residual variances.}

\item{\code{V}}{A list of V matrices, representing the prior variance for each component.}

\item{\code{pi}}{A numeric vector of prior weights for each component.}

\item{\code{estimate_residual_variance}}{A logical value indicating whether the residual variance is estimated (\code{TRUE}) or fixed (\code{FALSE}).}

\item{\code{L}}{An integer specifying the number of causal components in the model.}

\item{\code{nSNP}}{An integer representing the total number of SNPs analyzed.}

\item{\code{nsex}}{An integer representing the number of sexes analyzed (typically 2: male and female).}

\item{\code{Credible_Set}}{A list of credible sets, each representing SNP sets with high posterior probabilities of causality.}

\item{\code{PIP}}{A numeric vector of Posterior Inclusion Probabilities (PIP) for each SNP.}

\item{\code{PIP_config}}{A matrix of PIP values for each SNP across different configurations.}

\item{\code{posterior_mean}}{A matrix containing the posterior means for \eqn{\beta_c} (common effects) and \eqn{\beta_d} (sex-dimorphic effects).}

\item{\code{posterior_variance}}{A matrix with the first column being the variance for \eqn{\beta_c}, the second column being the variance for \eqn{\beta_d}, and the third column being the covariance between \eqn{\beta_c} and \eqn{\beta_d}.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{name_config}}{A character vector containing the names of causal configurations.}

\item{\code{column_config}}{A list specifying the possible configurations for columns.}

\item{\code{alpha}}{A list of alpha matrices, where each matrix corresponds to posterior inclusion probabilities for each component.}

\item{\code{mu1}}{A list of mu1 lists, where each list contains the posterior means for each causal configuration for each component.}

\item{\code{mu2}}{A list of mu2 lists, where each list contains the posterior variances for each causal configuration for each component.}

\item{\code{EB1}}{A list of EB1 matrices, where each matrix represents the posterior means of \eqn{\beta_c} and \eqn{\beta_d} for each component.}

\item{\code{EB2}}{A list of EB2 matrices, where each matrix represents the posterior variances for each component. The first column is the variance for \eqn{\beta_c}, the second column is the variance for \eqn{\beta_d}, and the third column is the covariance between \eqn{\beta_c} and \eqn{\beta_d}.}

\item{\code{Xr}}{A list of Xr matrices, representing the genetic part for each component.}

\item{\code{KL}}{A numeric vector containing the Kullback-Leibler (KL) divergence values for each component.}

\item{\code{lbf}}{A list of lbf matrices, where each matrix represents the log-Bayes factors for causal configuration for each component.}

\item{\code{ELBO}}{A numeric vector of Evidence Lower Bound (ELBO) values for each iteration, used to assess model convergence.}

\item{\code{sigma2}}{A numeric vector of residual variances.}

\item{\code{V}}{A list of V matrices, representing the prior variance for each component.}

\item{\code{pi}}{A numeric vector of prior weights for each component.}

\item{\code{estimate_residual_variance}}{A logical value indicating whether the residual variance is estimated (\code{TRUE}) or fixed (\code{FALSE}).}

\item{\code{L}}{An integer specifying the number of causal components in the model.}

\item{\code{nSNP}}{An integer representing the total number of SNPs analyzed.}

\item{\code{nsex}}{An integer representing the number of sexes analyzed (typically 2: male and female).}

\item{\code{Credible_Set}}{A list of credible sets, each representing SNP sets with high posterior probabilities of causality.}

\item{\code{PIP}}{A numeric vector of Posterior Inclusion Probabilities (PIP) for each SNP.}

\item{\code{PIP_config}}{A matrix of PIP values for each SNP across different configurations.}

\item{\code{posterior_mean}}{A matrix containing the posterior means for \eqn{\beta_c} (common effects) and \eqn{\beta_d} (sex-dimorphic effects).}

\item{\code{posterior_variance}}{A matrix with the first column being the variance for \eqn{\beta_c}, the second column being the variance for \eqn{\beta_d}, and the third column being the covariance between \eqn{\beta_c} and \eqn{\beta_d}.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-sdSuSiEObject-new}{\code{sdSuSiEObject$new()}}
\item \href{#method-sdSuSiEObject-compute_residual}{\code{sdSuSiEObject$compute_residual()}}
\item \href{#method-sdSuSiEObject-compute_Xr}{\code{sdSuSiEObject$compute_Xr()}}
\item \href{#method-sdSuSiEObject-par_update}{\code{sdSuSiEObject$par_update()}}
\item \href{#method-sdSuSiEObject-compute_SER_posterior_loglik}{\code{sdSuSiEObject$compute_SER_posterior_loglik()}}
\item \href{#method-sdSuSiEObject-compute_KL}{\code{sdSuSiEObject$compute_KL()}}
\item \href{#method-sdSuSiEObject-update_residual_variance}{\code{sdSuSiEObject$update_residual_variance()}}
\item \href{#method-sdSuSiEObject-get_result}{\code{sdSuSiEObject$get_result()}}
\item \href{#method-sdSuSiEObject-sdSuSiE_summary}{\code{sdSuSiEObject$sdSuSiE_summary()}}
\item \href{#method-sdSuSiEObject-clone}{\code{sdSuSiEObject$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sdSuSiEObject-new"></a>}}
\if{latex}{\out{\hypertarget{method-sdSuSiEObject-new}{}}}
\subsection{Method \code{new()}}{
Initialize the sdSuSiE object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sdSuSiEObject$new(
  p,
  L,
  N_sex,
  residual_variance,
  prior_weights,
  estimate_residual_variance,
  max_iter,
  name_vec
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{p}}{Integer, number of SNPs.}

\item{\code{L}}{Integer, number of causal components.}

\item{\code{N_sex}}{Integer, number of sexes analyzed (typically 2: male and female).}

\item{\code{residual_variance}}{Numeric vector of initial residual variances.}

\item{\code{prior_weights}}{Numeric vector of prior weights.}

\item{\code{estimate_residual_variance}}{Logical, if TRUE, estimate the residual variance.}

\item{\code{max_iter}}{Integer, maximum number of iterations.}

\item{\code{name_vec}}{Character vector, names for configurations.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `sdSuSieObject` object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sdSuSiEObject-compute_residual"></a>}}
\if{latex}{\out{\hypertarget{method-sdSuSiEObject-compute_residual}{}}}
\subsection{Method \code{compute_residual()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sdSuSiEObject$compute_residual(sdSuSiE_Data, sdSuSiE_Obj)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{sdSuSiE_Data}}{List, sdSuSiE data object.}

\item{\code{sdSuSiE_Obj, }}{current sdSuSiE object.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Matrix of computed residuals.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sdSuSiEObject-compute_Xr"></a>}}
\if{latex}{\out{\hypertarget{method-sdSuSiEObject-compute_Xr}{}}}
\subsection{Method \code{compute_Xr()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sdSuSiEObject$compute_Xr(sdSuSiE_Data, b1b2, l_index)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{sdSuSiE_Data}}{List, sdSuSiE data object.}

\item{\code{b1b2}}{Matrix, current estimate of b.}

\item{\code{l_index}}{Integer, index of the current component.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Updated Xr matrix for the specified component.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sdSuSiEObject-par_update"></a>}}
\if{latex}{\out{\hypertarget{method-sdSuSiEObject-par_update}{}}}
\subsection{Method \code{par_update()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sdSuSiEObject$par_update(SER_res, l_index)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{SER_res}}{List, results from the SER analysis.}

\item{\code{l_index}}{Integer, index for the current component.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Updated `sdSuSiEObject` object containing the analysis results and model parameters.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sdSuSiEObject-compute_SER_posterior_loglik"></a>}}
\if{latex}{\out{\hypertarget{method-sdSuSiEObject-compute_SER_posterior_loglik}{}}}
\subsection{Method \code{compute_SER_posterior_loglik()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sdSuSiEObject$compute_SER_posterior_loglik(sdSuSiE_Data, comp_residual, b1b2)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{sdSuSiE_Data}}{List, sdSuSiE data object.}

\item{\code{comp_residual}}{Matrix, residual values for the current component.}

\item{\code{b1b2}}{List, containing the EB1 and EB2 matrices.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Log-likelihood value for the SER posterior.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sdSuSiEObject-compute_KL"></a>}}
\if{latex}{\out{\hypertarget{method-sdSuSiEObject-compute_KL}{}}}
\subsection{Method \code{compute_KL()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sdSuSiEObject$compute_KL(SER_res, value, l_index)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{SER_res}}{List, results from the SER analysis.}

\item{\code{value}}{Numeric, value to compute KL divergence.}

\item{\code{l_index}}{Integer, index of the current component.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Updated KL divergence value for the specified component.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sdSuSiEObject-update_residual_variance"></a>}}
\if{latex}{\out{\hypertarget{method-sdSuSiEObject-update_residual_variance}{}}}
\subsection{Method \code{update_residual_variance()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sdSuSiEObject$update_residual_variance(sdSuSiE_Data, niter)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{sdSuSiE_Data}}{List, sdSuSiE data object.}

\item{\code{niter}}{Integer, the current iteration number.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A numeric vector with the updated residual variances.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sdSuSiEObject-get_result"></a>}}
\if{latex}{\out{\hypertarget{method-sdSuSiEObject-get_result}{}}}
\subsection{Method \code{get_result()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sdSuSiEObject$get_result(cs, pip, pip_config)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{cs}}{List, credible sets for effects.}

\item{\code{pip}}{Numeric vector, posterior inclusion probabilities for each SNP.}

\item{\code{pip_config}}{Matrix, PIP values for each SNP across}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sdSuSiEObject-sdSuSiE_summary"></a>}}
\if{latex}{\out{\hypertarget{method-sdSuSiEObject-sdSuSiE_summary}{}}}
\subsection{Method \code{sdSuSiE_summary()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sdSuSiEObject$sdSuSiE_summary(sdSuSiE_Data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{sdSuSiE_Data}}{List, sdSuSiE data object.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sdSuSiEObject-clone"></a>}}
\if{latex}{\out{\hypertarget{method-sdSuSiEObject-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sdSuSiEObject$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
