% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdSuSiE.R
\name{single_effect_regression}
\alias{single_effect_regression}
\title{The function for single effect regression in sdSuSiE.}
\usage{
single_effect_regression(XtR, XtX.diag, sdSuSiEObject_obj, l_index)
}
\arguments{
\item{XtR}{A list of Xr matrices, representing the genetic part for the specific component.}

\item{XtX.diag}{A numeric vector containing the diagonal elements of the XtX matrix.}

\item{sdSuSiEObject_obj}{An R6 object for handling the sdSuSiE analysis and storing its results.}

\item{l_index}{A numeric value specifying the index of the component for which single-effect regression is performed.}
}
\value{
A list with alphas, posteriors, likelihood, prior variance estimates, for single effect regression result for the specific component. 
\item{alpha}{A matrix of dimension \eqn{p \times k}, where \eqn{p} is the number of SNPs, and \eqn{k} is the number of causal configurations. Each column contains the PIPs for a specific causal scenario.}
\item{mu1_multi}{A list of posterior means for each causal configuration.}
\item{mu2_multi}{A list of posterior variances for each causal configuration.}
\item{lbf_multi}{A matrix containing log-Bayes factors for each causal configuration.}
\item{V}{A 2 by 2 matrix representing the prior variance estimates.}
\item{loglik}{The logarithm of the likelihood.} 
\item{b1b2}{A list containing EB1 and EB2 matrices.}
}
\description{
single_effect_regression conducts single effect regression for the specific component.
}
