% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdSuSiE.R
\name{vec_to_cov}
\alias{vec_to_cov}
\title{The function transforms the estimated parameter vector into a covariance matrix.}
\usage{
vec_to_cov(v_vec, diag_index, nsex)
}
\arguments{
\item{v_vec}{The estimated parameter vector.}

\item{diag_index}{Index of the variance in the parameter vector.}

\item{nsex}{Upper bound of the prior variance.}
}
\value{
A constructed covariance matrix.
}
\description{
The function transforms the estimated parameter vector into a covariance matrix.
}
